requirements.txt
streamlit
langgraph
langchain-groq
python-dotenv
arxiv
requests
beautifulsoup4
plotly
pinecone
sentence-transformers
